third_party/WebKit/LayoutTests/http/tests/websocket/long-control-frame-expected.txt

CONSOLE ERROR: line 15: WebSocket connection to 'ws://localhost:8880/long-control-frame' failed: Invalid frame header
Test whether WebSocket rejects control frames longer than 125 bytes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
